refactor(utils): split _utils.py into _io/_contigs/_genos by domain
