+++ python/genoray/_var_ranges.py
@@ -11,8 +11,9 @@ import polars_config_meta  # noqa: F401
 from numpy.typing import ArrayLike, NDArray
 from seqpro.rag import OFFSET_TYPE, lengths_to_offsets
 
+from ._contigs import ContigNormalizer
 from ._types import POS_TYPE, V_IDX_TYPE
-from ._utils import DTYPE, ContigNormalizer, np_to_pl_dtype
+from ._utils import DTYPE, np_to_pl_dtype
 
 _logging.getLogger("polars_bio").setLevel(_logging.ERROR)
