+++ tests/test_gtf_annotation.py
@@ -12,7 +12,7 @@ import pytest
 from pytest_cases import case, parametrize_with_cases
 
 from genoray._svar._annotate import _empty_annot, _get_strand_and_codon_pos
-from genoray._utils import ContigNormalizer
+from genoray._contigs import ContigNormalizer
 
 GTF_SCHEMA = {
     "chrom": pl.Utf8,
